move mobile vitals integration to own file
