+++ flutter/lib/src/default_integrations.dart
@@ -1,14 +1,12 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:sentry/sentry.dart';
 
 import 'sentry_flutter_options.dart';
-import 'sentry_native_wrapper.dart';
 import 'widgets_binding_observer.dart';
 
 /// It is necessary to initialize Flutter method channels so that our plugin can
@@ -509,68 +507,5 @@ class LoadReleaseIntegration extends Integration<SentryFlutterOptions> {
         .replaceAll('\n', '_')
         // replace Unicode NULL character with an empty string
         .replaceAll('\u{0000}', '');
-  }
-}
-
-/// Enrich [SentryTransaction] objects with native app data for mobile vitals.
-class MobileVitalsIntegration extends Integration<SentryFlutterOptions> {
-  MobileVitalsIntegration(this._nativeWrapper);
-
-  final SentryNativeWrapper _nativeWrapper;
-
-  @override
-  FutureOr<void> call(Hub hub, SentryFlutterOptions options) {
-    if (options.autoAppStart) {
-      SchedulerBinding.instance?.addPostFrameCallback((timeStamp) {
-        options.appStartFinish = DateTime.now();
-      });
-    }
-
-    options.addEventProcessor(
-        _NativeAppStartEventProcessor(_nativeWrapper, options));
-
-    options.sdk.addIntegration('mobileVitalsIntegration');
-  }
-}
-
-class _NativeAppStartEventProcessor extends EventProcessor {
-  _NativeAppStartEventProcessor(this._nativeWrapper, this._options);
-
-  final SentryNativeWrapper _nativeWrapper;
-  final SentryFlutterOptions _options;
-
-  var _didFetchAppStart = false;
-
-  @override
-  FutureOr<SentryEvent?> apply(SentryEvent event, {hint}) async {
-    final appStartFinishTime = _options.appStartFinish;
-
-    if (appStartFinishTime != null &&
-        event is SentryTransaction &&
-        !_didFetchAppStart) {
-      _didFetchAppStart = true;
-
-      final nativeAppStart = await _nativeWrapper.fetchNativeAppStart();
-      if (nativeAppStart == null) {
-        return event;
-      } else {
-        return event.copyWith(
-            measurements: [nativeAppStart.toMeasurement(appStartFinishTime)]);
-      }
-    } else {
-      return event;
-    }
-  }
-}
-
-extension NativeAppStartMeasurement on NativeAppStart {
-  SentryMeasurement toMeasurement(DateTime appStartFinishTime) {
-    final appStartDateTime =
-        DateTime.fromMillisecondsSinceEpoch(appStartTime.toInt());
-    final duration = appStartFinishTime.difference(appStartDateTime);
-
-    return isColdStart
-        ? SentryMeasurement.coldAppStart(duration)
-        : SentryMeasurement.warmAppStart(duration);
   }
 }
