+++ flutter/test/default_integrations_test.dart
@@ -5,8 +5,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
-import 'package:sentry_flutter/src/sentry_native_wrapper.dart';
-import 'package:sentry/src/sentry_tracer.dart';
 
 import 'mocks.dart';
 import 'mocks.mocks.dart';
@@ -418,76 +416,14 @@ void main() {
       expect(fixture.options.dist, isNull);
     });
   });
-
-  group('$MobileVitalsIntegration', () {
-    late Fixture fixture;
-
-    setUp(() {
-      fixture = Fixture();
-    });
-
-    test('native app start measurement added to first transaction', () async {
-      fixture.options.autoAppStart = false;
-      fixture.options.appStartFinish = DateTime.fromMillisecondsSinceEpoch(10);
-      fixture.wrapper.nativeAppStart = NativeAppStart(0, true);
-
-      fixture.getMobileVitalsIntegration().call(MockHub(), fixture.options);
-
-      final tracer = fixture.createTracer();
-      final transaction = SentryTransaction(tracer);
-
-      final processor = fixture.options.eventProcessors.first;
-      final enriched = await processor.apply(transaction) as SentryTransaction;
-
-      final expected = SentryMeasurement('app_start_cold', 10);
-      expect(enriched.measurements?[0].name, expected.name);
-      expect(enriched.measurements?[0].value, expected.value);
-    });
-
-    test('native app start measurement not added to following transactions',
-        () async {
-      fixture.options.autoAppStart = false;
-      fixture.options.appStartFinish = DateTime.fromMillisecondsSinceEpoch(10);
-      fixture.wrapper.nativeAppStart = NativeAppStart(0, true);
-
-      fixture.getMobileVitalsIntegration().call(MockHub(), fixture.options);
-
-      final tracer = fixture.createTracer();
-      final transaction = SentryTransaction(tracer);
-
-      final processor = fixture.options.eventProcessors.first;
-
-      var enriched = await processor.apply(transaction) as SentryTransaction;
-      var secondEnriched = await processor.apply(enriched) as SentryTransaction;
-
-      expect(secondEnriched.measurements?.length, 1);
-    });
-  });
 }
 
 class Fixture {
   final hub = MockHub();
-  final wrapper = MockNativeWrapper();
   final options = SentryFlutterOptions(dsn: fakeDsn);
 
   LoadReleaseIntegration getIntegration({PackageLoader? loader}) {
     return LoadReleaseIntegration(loader ?? loadRelease);
-  }
-
-  MobileVitalsIntegration getMobileVitalsIntegration() {
-    return MobileVitalsIntegration(wrapper);
-  }
-
-  // ignore: invalid_use_of_internal_member
-  SentryTracer createTracer({
-    bool? sampled,
-  }) {
-    final context = SentryTransactionContext(
-      'name',
-      'op',
-      sampled: sampled,
-    );
-    return SentryTracer(context, MockHub());
   }
 
   Future<PackageInfo> loadRelease() {
